requirements.txt
openai>=1.0.0
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=0.19.0
jupyter>=1.0.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0 